refactor: result 관련 클래스를 패키지로 묶음
