+++ src/main/java/blackjack/controller/BlackJackController.java
@@ -2,7 +2,7 @@ package blackjack.controller;
 
 import blackjack.domain.card.CardDeck;
 import blackjack.domain.user.Dealer;
-import blackjack.domain.user.DealerResult;
+import blackjack.domain.result.DealerResult;
 import blackjack.domain.user.Player;
 import blackjack.domain.user.Players;
 import blackjack.view.InputView;
