+++ src/test/java/blackjack/domain/user/DealerResultTest.java
@@ -6,6 +6,7 @@ import blackjack.domain.card.Card;
 import blackjack.domain.card.CardNumber;
 import blackjack.domain.card.CardSymbol;
 import blackjack.domain.card.UserDeck;
+import blackjack.domain.result.DealerResult;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
